Change namespace to Resque::Plugins and remove Lock from namespace
