+++ lib/resque/plugins/uniqueness/none.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module Resque
-  module Uniqueness
-    module Lock
+  module Plugins
+    module Uniqueness
       # Uses for cases when plugin not included for certain job
       # or when user want to disable uniqueness for certain job
       #   @lock_type = :none
