+++ lib/resque/plugins/uniqueness/version.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Resque
+  module Plugins
+    module Uniqueness
+      VERSION = '0.0.1'
+    end
+  end
+end
