--- lib/resque/uniqueness/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Resque
-  module Uniqueness
-    VERSION = '0.0.1'
-  end
-end
