+++ spec/resque/plugins/uniqueness/job_extension_spec.rb
@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-require_relative '../../shared_contexts/with_lock_spec'
+require_relative '../../../shared_contexts/with_lock_spec'
 
 # We already prepended this module in `lib/resque/uniqueness.rb`
 # Therefore we will test Resque::Job class
-RSpec.describe Resque::Uniqueness::JobExtension do
+RSpec.describe Resque::Plugins::Uniqueness::JobExtension do
   describe '.create' do
     subject { Resque::Job.create(queue, klass, *args) }
 
     include_context 'with lock', :locked_on_schedule, :try_lock_schedule
-    let(:lock_class) { Resque::Uniqueness::Lock::UntilAndWhileExecuting }
+    let(:lock_class) { Resque::Plugins::Uniqueness::UntilAndWhileExecuting }
     let(:queue) { :test_queue }
     let(:klass) { UntilAndWhileExecutingWorker }
     let(:args) { [] }
@@ -53,13 +53,13 @@ RSpec.describe Resque::Uniqueness::JobExtension do
     subject { Resque::Job.reserve(queue) }
 
     include_context 'with lock', :ensure_unlock_schedule, :try_lock_perform
-    let(:lock_class) { Resque::Uniqueness::Lock::UntilAndWhileExecuting }
+    let(:lock_class) { Resque::Plugins::Uniqueness::UntilAndWhileExecuting }
     let(:queue) { :test_queue }
     let(:job) { Resque::Job.new(queue, job_payload) }
     let(:job_payload) { {'class' => UntilAndWhileExecutingWorker, 'args' => args} }
     let(:args) { [] }
 
-    before { allow(Resque::Uniqueness).to receive(:pop_perform_unlocked_from_queue).and_return(job) }
+    before { allow(Resque::Plugins::Uniqueness).to receive(:pop_perform_unlocked_from_queue).and_return(job) }
 
     its_block { is_expected.to send_message(lock_instance, :ensure_unlock_schedule) }
     its_block { is_expected.to send_message(lock_instance, :try_lock_perform) }
@@ -90,11 +90,11 @@ RSpec.describe Resque::Uniqueness::JobExtension do
 
     before do
       allow(Resque::Job).to receive(:data_store).and_return(data_store_instance)
-      allow(Resque::Uniqueness).to receive(:destroy)
+      allow(Resque::Plugins::Uniqueness).to receive(:destroy)
     end
 
     its_block { is_expected.to send_message(data_store_instance, :remove_from_queue).returning(2) }
-    its_block { is_expected.to send_message(Resque::Uniqueness, :destroy) }
+    its_block { is_expected.to send_message(Resque::Plugins::Uniqueness, :destroy) }
     it { is_expected.to eq 2 }
 
     context 'when resque inline' do
@@ -105,7 +105,7 @@ RSpec.describe Resque::Uniqueness::JobExtension do
       end
 
       its_block { is_expected.to send_message(data_store_instance, :remove_from_queue).returning(2) }
-      its_block { is_expected.not_to send_message(Resque::Uniqueness, :destroy) }
+      its_block { is_expected.not_to send_message(Resque::Plugins::Uniqueness, :destroy) }
       it { is_expected.to eq 2 }
     end
   end
